- Replaced boring_checkboxgroup_field with checkbox boring_checkbox_field - Changed the class for only one checkbox - Fix JSON KEY didn't show correct value on form controller
